+++ retina/__init__.py
@@ -34,7 +34,7 @@ from .detect import (
     VlmDetector,
     YoloDetector,
 )
-from .embed import DinoV2Embedder
+from .embed import DinoV2Embedder, VJepa2Embedder
 from .eval import event_f1, match_events
 from .events import SPEC, Event, EventType, Frame
 from .export import JsonlSink, WebhookSink, to_jsonl
@@ -87,6 +87,7 @@ __all__ = [
     "NorfairTracker",
     # latent producers
     "DinoV2Embedder",
+    "VJepa2Embedder",
     # zones / rules
     "Zone",
     "Line",